backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
google-cloud-firestore[async]>=2.16.0
firebase-admin>=6.0.0
google-genai>=1.0.0
google-cloud-storage>=2.18.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27
